LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/bufferedAmount/bufferedAmount-deleting-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://web-platform.test:49001/echo' failed: WebSocket is closed before the connection is established.

PASS WebSockets: delete bufferedAmount
